Move CQLDetectWithin to a new module
