+++ lib/query.py
@@ -1,0 +1,118 @@
+# Copyright (c) 2017 Charles University in Prague, Faculty of Arts,
+#                    Institute of the Czech National Corpus
+# Copyright (c) 2017 Tomas Machalek <tomas.machalek@gmail.com>
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; version 2
+# dated June, 1991.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+import re
+
+
+class CQLDetectWithin(object):
+    """
+    A simplified parser to detect 'within' part of a CQL query.
+    """
+
+    @staticmethod
+    def split_by_parentheses(s):
+        if type(s) is tuple:
+            return [s]
+        return [v1 for v2 in [re.split(r'(\])', x) for x in re.split(r'(\[)', s)] for v1 in v2]
+
+    @staticmethod
+    def split_by_whitespace(s):
+        items = re.split(r'\s+', s)
+        return [item for sublist in zip(items, len(items) * [' ']) for item in sublist][:-1]
+
+    def parse_lex_elems(self, s):
+        i = 0
+        ans = []
+        curr_piece = ''
+        state = 0   # '1' = opened ", '2' = opened '
+        while i < len(s):
+            if s[i] == '\\':
+                curr_piece += s[i+1]
+                i += 2
+                continue
+            if s[i] == '"':
+                if state == 0:
+                    ans.extend(self.split_by_whitespace(curr_piece))
+                    curr_piece = ''
+                    state = 1
+                elif state == 1:
+                    ans.append((curr_piece, 's'))  # tuple => string part of the query
+                    curr_piece = ''
+                    state = 0
+                else:
+                    raise Exception('syntax error')
+            else:
+                curr_piece += s[i]
+            i += 1
+        if len(curr_piece) > 0:
+            ans.extend(self.split_by_whitespace(curr_piece))
+        return ans
+
+    @staticmethod
+    def get_next_token(struct, start):
+        for i in range(start + 1, len(struct)):
+            if len(struct[i].strip()) > 0:
+                return struct[i]
+        return None
+
+    @staticmethod
+    def next_token_matches(struct, start, pattern):
+        token = CQLDetectWithin.get_next_token(struct, start)
+        if token:
+            return re.match(pattern, token) is not None
+        return False
+
+    def get_within_part(self, s):
+        def join_items(contains, items, idx):
+            if idx is None:
+                return None
+            ans = []
+            for item in items[idx:]:
+                if type(item) is tuple:
+                    ans.append(u'"%s"' % (item[0],))
+                else:
+                    ans.append(item)
+            return u''.join(ans)
+
+        return self.analyze(s, join_items)
+
+    def analyze(self, s, on_hit):
+        if type(s) in (str, unicode):
+            struct = self.parse(s)
+        else:
+            struct = s
+        last_p = None
+        for i, item in enumerate(struct):
+            if item is None:
+                continue
+            if item in (']', '['):
+                last_p = item
+            elif item in ('within', '!within', 'within!'):
+                if self.next_token_matches(struct, i, r'\w+:'):
+                    pass  # aligned corpus search
+                elif self.next_token_matches(struct, i, r'<.+'):
+                    return on_hit(True, struct, i)
+                elif last_p in (']', None):
+                    return on_hit(True, struct, i)
+        return on_hit(False, struct, None)
+
+    def parse(self, s):
+        result = []
+        ans = self.parse_lex_elems(s)
+        for item in ans:
+            x = self.split_by_parentheses(item)
+            result.extend(x)
+        result = [x for x in result if x != '']
+        return result
+
